requirements.txt
cffi==1.17.1
cryptography==45.0.4
mysql-connector-python==9.3.0
pycparser==2.22
